% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_gdd.R
\name{calculate_gdd}
\alias{calculate_gdd}
\title{Calculate growing degree days}
\usage{
calculate_gdd(temp_min, temp_max, base_temp = 5)
}
\arguments{
\item{temp_min}{Vector of minimum daily temperatures}

\item{temp_max}{Vector of maximum daily temperatures}

\item{base_temp}{Base temperature threshold (default 5°C)}
}
\value{
Vector of growing degree days
}
\description{
Calculate growing degree days
}
\examples{
calculate_gdd(c(10, 12), c(20, 22), base_temp = 5)
}
