% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genotydata}
\alias{genotydata}
\title{MOCK Genotypic and molecular data from oak holobiont}
\format{
A data frame with 50 rows and 12 variables:
\describe{
\item{tree_id}{Tree identifier}
\item{site_id}{Site identifier}
\item{snp_001 to snp_005}{SNP genotypes (0=homozygous ref, 1=heterozygous, 2=homozygous alt)}
\item{epigenetic_mark_1, epigenetic_mark_2}{Epigenetic modification levels}
\item{gene_expression_drought}{Drought response gene expression level}
\item{gene_expression_herbivory}{Herbivory response gene expression level}
\item{microbiome_diversity}{Shannon diversity index of associated microbiome}
}
}
\source{
PhytOakmeter research unit
}
\usage{
genotydata
}
\description{
Multi-omics data including SNPs, epigenetic marks, gene expression, and
microbiome diversity from Quercus robur clone DF159.
}
\keyword{datasets}
